refactor examples/alltoall.c into a helpers header and rename main program to alltoall_demo.c
